Fixed the issue of a non-responsive GPIO
What happened was that I created a variable, floatSwitch, and I assigned it to
the GPIO value. I then called that variable later in the program. I expected
the variable to check the GPIO again, but in reality, it was actually just
regurgitating the same value that it was assigned at the beining of the program.